Handing in project
Overall site design is finished. For demo day, I want to add more
events and articles, and two of the articles still need content (they
currently link to outside sources). I also still need to do the final
edit of the background video and upload it to Vimeo for optimized
streaming.